+++ example/extension_hats/hat_light_sensor.dart
@@ -13,7 +13,7 @@ const wait = 500;
 ///
 /// Usage: [nano|grove|grovePlus] analogPin
 void main(List<String> args) {
-  var tupple = checkArgs(args);
+  var tupple = checkArgs(args, "analogPin");
   var pin = tupple.$2;
   switch (tupple.$1) {
     case Hat.nano:
@@ -43,7 +43,7 @@ void main(List<String> args) {
 
       while (true) {
         print(hat.readADCraw(pin));
-        sleep(Duration(milliseconds: 100));
+        sleep(Duration(milliseconds: wait));
       }
   }
 }